Relation模块更新

+++ server/relation/db/table.go
@@ -2,6 +2,6 @@ package db
 
 type RelationDb struct {
 	Id         int64 `gorm:"primarykey;auto_increment"`
-	Userid     int   `gorm:"not null"`
-	Followerid int   `gorm:"not null"`
+	Userid     int64 `gorm:"not null"`
+	Followerid int64 `gorm:"not null"`
 }
